adjust merge
